Move emoji utilities to non-React file
